backend/requirements.txt
# ── Web framework ──────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0

# ── Base de dados ──────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# ── Autenticação ───────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# ── Configuração ───────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Scraping ───────────────────────────────────────────────
playwright==1.44.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1

# ── Utilitários ────────────────────────────────────────────
tenacity==8.3.0          # retry logic para scrapers
schedule==1.2.1          # jobs agendados simples
python-dateutil==2.9.0

# ── Dev / testes ───────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0            # cliente de testes FastAPI
